+++ changelog.markdown
@@ -1,3 +1,11 @@
+## 1.48.0.0
+
+* Bundled with docu.exe compiled against .Net 4.0.
+* Fixed docu calls to run with full filenames.
+* Added targetplatform, target and log switches for ILMerge task.
+* Added Git.Information.getLastTag() which gets the last git tag by calling git describe.
+* Added Git.Information.getCurrentHash() which gets the last current sha1.
+
 ## 1.46.0.0
 
 * Fixed Nuget support and allows automatic push.